Fix dart2js Uint32x4List in checked mode
TBR

Review URL: https://codereview.chromium.org//48783003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@29640 260f80e4-7a28-3924-810f-c04153c831b5

+++ sdk/lib/typed_data/dart2js/typed_data_dart2js.dart
@@ -870,10 +870,10 @@ class Uint32x4List
 
   Uint32x4 operator[](int index) {
     _checkIndex(index, length);
-    double _x = _storage[(index*4)+0];
-    double _y = _storage[(index*4)+1];
-    double _z = _storage[(index*4)+2];
-    double _w = _storage[(index*4)+3];
+    int _x = _storage[(index*4)+0];
+    int _y = _storage[(index*4)+1];
+    int _z = _storage[(index*4)+2];
+    int _w = _storage[(index*4)+3];
     return new Uint32x4(_x, _y, _z, _w);
   }
